habanalabs: Elide a warning on 32-bit targets
This double-cast pattern looks a bit awkward, but it already exists
elsewhere in the driver.  Without this patch I get

drivers/misc/habanalabs/common/memory.c: In function ‘alloc_device_memory’:
drivers/misc/habanalabs/common/memory.c:153:49: warning: cast from pointer to integer of different size [-Wpointer-to-int-cast]
  153 |                                                 (u64) gen_pool_dma_alloc_align(vm->dram_pg_pool,
      |                                                 ^

which ends up promoted to a build error in my test setup.

Fixes: e8458e20e0a3 ("habanalabs: make sure device mem alloc is page aligned")

+++ drivers/misc/habanalabs/common/memory.c
@@ -150,12 +150,12 @@ static int alloc_device_memory(struct hl_ctx *ctx, struct hl_mem_in *args,
 		for (i = 0 ; i < num_pgs ; i++) {
 			if (is_power_of_2(page_size))
 				phys_pg_pack->pages[i] =
-						(u64) gen_pool_dma_alloc_align(vm->dram_pg_pool,
-										page_size, NULL,
-										page_size);
+						(u64) (uintptr_t) gen_pool_dma_alloc_align(vm->dram_pg_pool,
+											   page_size, NULL,
+											   page_size);
 			else
-				phys_pg_pack->pages[i] = (u64) gen_pool_alloc(vm->dram_pg_pool,
-										page_size);
+				phys_pg_pack->pages[i] = (u64) (uintptr_t) gen_pool_alloc(vm->dram_pg_pool,
+											  page_size);
 			if (!phys_pg_pack->pages[i]) {
 				dev_err(hdev->dev,
 					"Failed to allocate device memory (out of memory)\n");
